Extract Store Options Types into options.rs.
Following on from refactoring types out of store.rs into types.rs, this
goes a bit further and extracts the common input option structs into
options.rs.

Next step is to add results.rs, which will contain the output types.
